+++ module/shell/src/main/clj/convex/shell/req/cell.clj
@@ -10,6 +10,17 @@
 ;;;;;;;;;;
 
 
+(defn ref-stat
+
+  "Requests for providing stats about the `cell`'s refs."
+
+  [ctx [cell]]
+
+  ($.cvm/result-set ctx
+                    ($.cell/any ($.std/ref-stat cell))))
+
+
+
 (defn size
 
   "Requests for getting the full memory size of a cell."
@@ -22,14 +33,3 @@
                              ($.cell/* "Cell to measure cannot be `nil`")))
       ($.cvm/result-set ctx
                         ($.cell/long ($.std/memory-size cell)))))
-
-
-
-(defn softness
-
-  "Requests for counting direct and soft references of a cell."
-
-  [ctx [cell]]
-
-  ($.cvm/result-set ctx
-                    ($.cell/any ($.std/softness cell))))